requirements.txt
audioread==3.0.1
certifi==2025.7.14
cffi @ file:///D:/bld/cffi_1725560649097/work
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
einops==0.8.1
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.33.4
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy @ file:///D:/bld/numpy_1747544591878/work/dist/numpy-2.2.6-cp310-cp310-win_amd64.whl#sha256=af66576ce916c795ffa81be9620c8138a3fda6e1713b2bf7775a630e47eca3a8
packaging==25.0
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
progressbar==2.5
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydub==0.25.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rotary-embedding-torch==0.8.8
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
soundfile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1737836266465/work
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tortoise-tts==3.0.0
tqdm==4.67.1
transformers==4.31.0
typeguard==4.4.4
typing_extensions==4.14.1
Unidecode==1.4.0
urllib3==2.5.0
